Ensure local apply of snapshot uses overlay mode.
When apply the local overlaybd.commit file to the snapshotter,
the Prepare must not operate in block/dev mode, or else we will
end up writing an overlaybd.commit file onto an ext4 filesystem
that is backed by another overlaybd.commit file.

We do this by passing a special label, instructing overlaybd-snapshotter
that this Prepare call is not to prepare a writable layer, but to
write the pieces of an image.